requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
decorator==5.2.1
distro==1.9.0
ecdsa==0.19.1
executing==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langsmith==0.4.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
openai==1.97.0
orjson==3.11.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-http-client==3.3.7
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
gradio==5.33.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
tenacity==9.1.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
zstandard==0.23.0
